requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
mangum==0.17.0
pypdf2~=3.0.1
python-dotenv~=1.0.1
google-generativeai
PyPDF2


docx2txt~=0.8
fitz~=0.0.1.dev2
